+++ src/main/scala/lexer/tokens.scala
@@ -51,12 +51,14 @@ case object MODULE extends KEYWORD
 case object NEW extends KEYWORD
 case object OTHERWISE extends KEYWORD
 case object PRIVATE extends DECL
+case object PROXY extends CONTROL
 case object RECORD extends DECL
 case object RECUR extends CONTROL
 case object REIFY extends CONTROL
 case object REFER extends KEYWORD
 case object REPEAT extends CONTROL
 case object SET extends CONTROL
+case object SYNC extends CONTROL
 case object THEN extends CONTROL
 case object THROW extends CONTROL
 case object TRAIT extends DECL
@@ -179,11 +181,13 @@ object KEYWORD_MAP {
     ("otherwise", OTHERWISE),
     ("priv", PRIVATE),
     ("private", PRIVATE),
+    ("proxy", PROXY),
     ("record", RECORD),
     ("recur", RECUR),
     ("reify", REIFY),
     ("repeat", REPEAT),
     ("set", SET),
+    ("sync", SYNC),
     ("then", THEN),
     ("throw", THROW),
     ("trait", TRAIT),