+++ src/main/scala/parser/ast/expressions/literals/LiteralExpression.scala
@@ -35,6 +35,7 @@ object LiteralExpression extends ExpressionParser {
       case fsl if fsl.isInstanceOf[FSTRING_LITERAL] =>
         FStringLiteral(tokens.consume(classOf[FSTRING_LITERAL]).value)
       case _ =>
+        println(s"@@ tokens${tokens}")
         throw InvalidExpression(tokens.nextToken())
     }
   }